refactor: minor clean up in int tests
